pyproject.toml
[project]
name = "microsoft-iq"
version = "0.1.0"
description = "Three demos showcasing Microsoft IQ: Foundry knowledge bases, Work intelligence, and Fabric pipeline agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Ricardo Cataldi" }]

dependencies = [
    "azure-identity>=1.17.0",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.22.0",
    "azure-ai-projects>=1.0.0b5",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
]
fabric = [
    "azure-mgmt-fabric>=1.0.0b1",
    "semantic-link-sempy>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
